readline-transform: update to 1.0
